Moved configuration.py & sons to /util
